refactor: 소수를 포함한 계산 추가로 인한 코드 리팩토링

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -17,7 +17,7 @@ public class CalculatorController {
     public void start(){
         String s = inputView.getUserInput();
         List<String> validNumber = calculatorService.findValidNumber(s);
-        long result = calculator.calculatePlusNumber(validNumber);
+        double result = calculator.calculatePlusNumber(validNumber);
         outputView.printResult(result);
     }
